REF: Renamed whois_response.py to raw_whois_response.py REF: Made timeout an argument, but with a default value
